+++ projects/storefrontlib/src/cms-structure/outlet/index.ts
@@ -1,5 +1,8 @@
+export * from './outlet-ref/outlet-ref.directive';
+export * from './outlet-ref/outlet-ref.module';
+export * from './outlet.directive';
+export * from './outlet.model';
 export * from './outlet.module';
 export * from './outlet.service';
-export * from './outlet.directive';
-export * from './outlet-ref/outlet-ref.module';
-export * from './outlet-ref/outlet-ref.directive';
+export * from './style-ref/style-ref.directive';
+export * from './style-ref/style-ref.module';
